fix bugs in example - fix the wrong dataset - fix the interface of evaluation and inference
