+++ examples/federated_learning/surface_defect_detection/training_worker/train.py
@@ -12,31 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
-
-import numpy as np
-import keras.preprocessing.image as img_preprocessing
-
 from sedna.datasources import TxtDataParse
 from sedna.common.config import Context, BaseConfig
 from sedna.core.federated_learning import FederatedLearning
 
-from interface import Estimator
-
-
-def image_process(line):
-    file_path, label = line.split(',')
-    original_dataset_url = (
-        BaseConfig.original_dataset_url or BaseConfig.train_dataset_url
-    )
-    root_path = os.path.dirname(original_dataset_url)
-    file_path = os.path.join(root_path, file_path)
-    img = img_preprocessing.load_img(file_path).resize((128, 128))
-    data = img_preprocessing.img_to_array(img) / 255.0
-    label = [0, 1] if int(label) == 0 else [1, 0]
-    data = np.array(data)
-    label = np.array(label)
-    return [data, label]
+from interface import Estimator, image_process
 
 
 def main():
